Revert "DeploymentDescriptorParser"
This reverts commit bd39818c532c6e9a80b1262410b384fc5fd0e225.
